requirements.txt
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
